move non-generated interfaces to separate class
too confusing to write manual interfaces in the same class since it's
not as obvious when in the generated section (like it is for generated
attacks, items, etc.)
