+++ src/battle/attack/Attack.java
@@ -3,6 +3,11 @@ package battle.attack;
 import battle.ActivePokemon;
 import battle.Battle;
 import battle.effect.Effect;
+import battle.effect.EffectInterfaces.ItemSwapperEffect;
+import battle.effect.EffectInterfaces.PassableEffect;
+import battle.effect.EffectInterfaces.PowderMove;
+import battle.effect.EffectInterfaces.SapHealthEffect;
+import battle.effect.EffectInterfaces.SwapOpponentEffect;
 import battle.effect.EffectNamesies;
 import battle.effect.InvokeEffect;
 import battle.effect.InvokeInterfaces.AdvantageMultiplierMove;
@@ -17,22 +22,17 @@ import battle.effect.InvokeInterfaces.CritStageEffect;
 import battle.effect.InvokeInterfaces.DefogRelease;
 import battle.effect.InvokeInterfaces.EffectBlockerEffect;
 import battle.effect.InvokeInterfaces.EffectChanceMultiplierEffect;
-import battle.effect.InvokeInterfaces.ItemSwapperEffect;
 import battle.effect.InvokeInterfaces.MurderEffect;
 import battle.effect.InvokeInterfaces.OpponentEndAttackEffect;
 import battle.effect.InvokeInterfaces.OpponentIgnoreStageEffect;
 import battle.effect.InvokeInterfaces.OpponentStatSwitchingEffect;
-import battle.effect.InvokeInterfaces.PassableEffect;
-import battle.effect.InvokeInterfaces.PowderMove;
 import battle.effect.InvokeInterfaces.PowerChangeEffect;
 import battle.effect.InvokeInterfaces.RapidSpinRelease;
-import battle.effect.InvokeInterfaces.SapHealthEffect;
 import battle.effect.InvokeInterfaces.SelfAttackBlocker;
 import battle.effect.InvokeInterfaces.SemiInvulnerableBypasser;
 import battle.effect.InvokeInterfaces.SleepyFightsterEffect;
 import battle.effect.InvokeInterfaces.StatSwitchingEffect;
 import battle.effect.InvokeInterfaces.StickyHoldEffect;
-import battle.effect.InvokeInterfaces.SwapOpponentEffect;
 import battle.effect.InvokeInterfaces.TargetSwapperEffect;
 import battle.effect.attack.FixedDamageMove;
 import battle.effect.attack.MultiStrikeMove;
