+++ src/battle/effect/pokemon/PokemonEffect.java
@@ -7,10 +7,14 @@ import battle.attack.AttackNamesies;
 import battle.attack.Move;
 import battle.attack.MoveType;
 import battle.effect.Effect;
+import battle.effect.EffectInterfaces.AttackSelectionSelfBlockerEffect;
+import battle.effect.EffectInterfaces.PassableEffect;
+import battle.effect.EffectInterfaces.PhysicalContactEffect;
+import battle.effect.EffectInterfaces.ProtectingEffect;
+import battle.effect.EffectInterfaces.SapHealthEffect;
 import battle.effect.InvokeInterfaces.AbsorbDamageEffect;
 import battle.effect.InvokeInterfaces.AlwaysCritEffect;
 import battle.effect.InvokeInterfaces.AttackSelectionEffect;
-import battle.effect.InvokeInterfaces.AttackSelectionSelfBlockerEffect;
 import battle.effect.InvokeInterfaces.BeforeTurnEffect;
 import battle.effect.InvokeInterfaces.BracingEffect;
 import battle.effect.InvokeInterfaces.ChangeAttackTypeEffect;
@@ -30,13 +34,9 @@ import battle.effect.InvokeInterfaces.ItemBlockerEffect;
 import battle.effect.InvokeInterfaces.LevitationEffect;
 import battle.effect.InvokeInterfaces.OpponentAccuracyBypassEffect;
 import battle.effect.InvokeInterfaces.OpponentTrappingEffect;
-import battle.effect.InvokeInterfaces.PassableEffect;
-import battle.effect.InvokeInterfaces.PhysicalContactEffect;
 import battle.effect.InvokeInterfaces.PowderBlocker;
 import battle.effect.InvokeInterfaces.PowerChangeEffect;
-import battle.effect.InvokeInterfaces.ProtectingEffect;
 import battle.effect.InvokeInterfaces.RapidSpinRelease;
-import battle.effect.InvokeInterfaces.SapHealthEffect;
 import battle.effect.InvokeInterfaces.SelfAttackBlocker;
 import battle.effect.InvokeInterfaces.SemiInvulnerableBypasser;
 import battle.effect.InvokeInterfaces.StageChangingEffect;
