+++ src/battle/effect/status/StatusCondition.java
@@ -5,11 +5,11 @@ import battle.Battle;
 import battle.attack.AttackNamesies;
 import battle.attack.MoveType;
 import battle.effect.Effect;
+import battle.effect.EffectInterfaces.SimpleStatModifyingEffect;
 import battle.effect.InvokeEffect;
 import battle.effect.InvokeInterfaces.BeforeTurnEffect;
 import battle.effect.InvokeInterfaces.EndTurnEffect;
 import battle.effect.InvokeInterfaces.OpponentStatusReceivedEffect;
-import battle.effect.InvokeInterfaces.SimpleStatModifyingEffect;
 import battle.effect.InvokeInterfaces.SleepyFightsterEffect;
 import battle.effect.InvokeInterfaces.StatusPreventionEffect;
 import battle.effect.InvokeInterfaces.StatusReceivedEffect;
