+++ src/battle/effect/team/TeamEffect.java
@@ -5,13 +5,13 @@ import battle.Battle;
 import battle.attack.Attack;
 import battle.attack.AttackNamesies;
 import battle.effect.Effect;
+import battle.effect.EffectInterfaces.SimpleStatModifyingEffect;
 import battle.effect.InvokeInterfaces.BarrierEffect;
 import battle.effect.InvokeInterfaces.CritBlockerEffect;
 import battle.effect.InvokeInterfaces.DefogRelease;
 import battle.effect.InvokeInterfaces.EndBattleEffect;
 import battle.effect.InvokeInterfaces.EntryEffect;
 import battle.effect.InvokeInterfaces.RapidSpinRelease;
-import battle.effect.InvokeInterfaces.SimpleStatModifyingEffect;
 import battle.effect.InvokeInterfaces.StatusPreventionEffect;
 import battle.effect.source.CastSource;
 import battle.effect.status.StatusNamesies;
