+++ src/item/Item.java
@@ -6,6 +6,10 @@ import battle.attack.AttackNamesies;
 import battle.attack.Move;
 import battle.attack.MoveCategory;
 import battle.attack.MoveType;
+import battle.effect.EffectInterfaces.EntryEndTurnEffect;
+import battle.effect.EffectInterfaces.ItemSwapperEffect;
+import battle.effect.EffectInterfaces.PhysicalContactEffect;
+import battle.effect.EffectInterfaces.SimpleStatModifyingEffect;
 import battle.effect.EffectNamesies;
 import battle.effect.InvokeInterfaces.ApplyDamageEffect;
 import battle.effect.InvokeInterfaces.AttackSelectionEffect;
@@ -16,18 +20,14 @@ import battle.effect.InvokeInterfaces.DefiniteEscape;
 import battle.effect.InvokeInterfaces.EffectReceivedEffect;
 import battle.effect.InvokeInterfaces.EndTurnEffect;
 import battle.effect.InvokeInterfaces.EntryEffect;
-import battle.effect.InvokeInterfaces.EntryEndTurnEffect;
 import battle.effect.InvokeInterfaces.GroundedEffect;
 import battle.effect.InvokeInterfaces.HalfWeightEffect;
-import battle.effect.InvokeInterfaces.ItemSwapperEffect;
 import battle.effect.InvokeInterfaces.LevitationEffect;
 import battle.effect.InvokeInterfaces.OpponentApplyDamageEffect;
 import battle.effect.InvokeInterfaces.OpponentTakeDamageEffect;
-import battle.effect.InvokeInterfaces.PhysicalContactEffect;
 import battle.effect.InvokeInterfaces.PowderBlocker;
 import battle.effect.InvokeInterfaces.PowerChangeEffect;
 import battle.effect.InvokeInterfaces.RepellingEffect;
-import battle.effect.InvokeInterfaces.SimpleStatModifyingEffect;
 import battle.effect.InvokeInterfaces.StallingEffect;
 import battle.effect.InvokeInterfaces.StatProtectingEffect;
 import battle.effect.InvokeInterfaces.StrikeFirstEffect;
