+++ src/item/hold/PowerItem.java
@@ -1,6 +1,6 @@
 package item.hold;
 
-import battle.effect.InvokeInterfaces.SimpleStatModifyingEffect;
+import battle.effect.EffectInterfaces.SimpleStatModifyingEffect;
 import pokemon.Stat;
 
 public interface PowerItem extends EVItem, SimpleStatModifyingEffect {
