+++ src/pokemon/ability/Ability.java
@@ -7,6 +7,14 @@ import battle.attack.AttackNamesies;
 import battle.attack.Move;
 import battle.attack.MoveCategory;
 import battle.attack.MoveType;
+import battle.effect.EffectInterfaces.EntryEndTurnEffect;
+import battle.effect.EffectInterfaces.ItemSwapperEffect;
+import battle.effect.EffectInterfaces.MaxLevelWildEncounterEffect;
+import battle.effect.EffectInterfaces.PhysicalContactEffect;
+import battle.effect.EffectInterfaces.RepelLowLevelEncounterEffect;
+import battle.effect.EffectInterfaces.SimpleStatModifyingEffect;
+import battle.effect.EffectInterfaces.SwapOpponentEffect;
+import battle.effect.EffectInterfaces.TypedWildEncounterSelector;
 import battle.effect.EffectNamesies;
 import battle.effect.InvokeEffect;
 import battle.effect.InvokeInterfaces.AbsorbDamageEffect;
@@ -30,12 +38,9 @@ import battle.effect.InvokeInterfaces.EncounterRateMultiplier;
 import battle.effect.InvokeInterfaces.EndBattleEffect;
 import battle.effect.InvokeInterfaces.EndTurnEffect;
 import battle.effect.InvokeInterfaces.EntryEffect;
-import battle.effect.InvokeInterfaces.EntryEndTurnEffect;
 import battle.effect.InvokeInterfaces.HalfWeightEffect;
 import battle.effect.InvokeInterfaces.ItemBlockerEffect;
-import battle.effect.InvokeInterfaces.ItemSwapperEffect;
 import battle.effect.InvokeInterfaces.LevitationEffect;
-import battle.effect.InvokeInterfaces.MaxLevelWildEncounterEffect;
 import battle.effect.InvokeInterfaces.ModifyStageValueEffect;
 import battle.effect.InvokeInterfaces.MurderEffect;
 import battle.effect.InvokeInterfaces.NameChanger;
@@ -47,14 +52,11 @@ import battle.effect.InvokeInterfaces.OpponentPowerChangeEffect;
 import battle.effect.InvokeInterfaces.OpponentStatusReceivedEffect;
 import battle.effect.InvokeInterfaces.OpponentTakeDamageEffect;
 import battle.effect.InvokeInterfaces.OpponentTrappingEffect;
-import battle.effect.InvokeInterfaces.PhysicalContactEffect;
 import battle.effect.InvokeInterfaces.PowderBlocker;
 import battle.effect.InvokeInterfaces.PowerChangeEffect;
 import battle.effect.InvokeInterfaces.PriorityChangeEffect;
-import battle.effect.InvokeInterfaces.RepelLowLevelEncounterEffect;
 import battle.effect.InvokeInterfaces.SelfAttackBlocker;
 import battle.effect.InvokeInterfaces.SemiInvulnerableBypasser;
-import battle.effect.InvokeInterfaces.SimpleStatModifyingEffect;
 import battle.effect.InvokeInterfaces.SleepyFightsterEffect;
 import battle.effect.InvokeInterfaces.StageChangingEffect;
 import battle.effect.InvokeInterfaces.StallingEffect;
@@ -65,11 +67,9 @@ import battle.effect.InvokeInterfaces.StatusPreventionEffect;
 import battle.effect.InvokeInterfaces.StatusReceivedEffect;
 import battle.effect.InvokeInterfaces.StickyHoldEffect;
 import battle.effect.InvokeInterfaces.SuperDuperEndTurnEffect;
-import battle.effect.InvokeInterfaces.SwapOpponentEffect;
 import battle.effect.InvokeInterfaces.SwitchOutEffect;
 import battle.effect.InvokeInterfaces.TakeDamageEffect;
 import battle.effect.InvokeInterfaces.TargetSwapperEffect;
-import battle.effect.InvokeInterfaces.TypedWildEncounterSelector;
 import battle.effect.InvokeInterfaces.WeatherBlockerEffect;
 import battle.effect.InvokeInterfaces.WeatherEliminatingEffect;
 import battle.effect.InvokeInterfaces.WildEncounterAlterer;
