+++ src/test/ClassTest.java
@@ -2,6 +2,8 @@ package test;
 
 import battle.attack.Attack;
 import battle.attack.AttackInterface;
+import battle.effect.EffectInterfaces.EffectReleaser;
+import battle.effect.EffectInterfaces.PassableEffect;
 import battle.effect.InvokeInterfaces.AbsorbDamageEffect;
 import battle.effect.InvokeInterfaces.AlwaysCritEffect;
 import battle.effect.InvokeInterfaces.ApplyDamageEffect;
@@ -23,7 +25,6 @@ import battle.effect.InvokeInterfaces.DefendingNoAdvantageChanger;
 import battle.effect.InvokeInterfaces.DefogRelease;
 import battle.effect.InvokeInterfaces.DifferentStatEffect;
 import battle.effect.InvokeInterfaces.EffectBlockerEffect;
-import battle.effect.InvokeInterfaces.EffectReleaser;
 import battle.effect.InvokeInterfaces.EncounterRateMultiplier;
 import battle.effect.InvokeInterfaces.EndBattleEffect;
 import battle.effect.InvokeInterfaces.EndTurnEffect;
@@ -44,7 +45,6 @@ import battle.effect.InvokeInterfaces.OpponentStatSwitchingEffect;
 import battle.effect.InvokeInterfaces.OpponentStatusReceivedEffect;
 import battle.effect.InvokeInterfaces.OpponentTakeDamageEffect;
 import battle.effect.InvokeInterfaces.OpponentTrappingEffect;
-import battle.effect.InvokeInterfaces.PassableEffect;
 import battle.effect.InvokeInterfaces.PowerChangeEffect;
 import battle.effect.InvokeInterfaces.PriorityChangeEffect;
 import battle.effect.InvokeInterfaces.RapidSpinRelease;
