+++ src/test/battle/AttackTest.java
@@ -6,10 +6,10 @@ import battle.attack.Move;
 import battle.attack.MoveCategory;
 import battle.attack.MoveType;
 import battle.effect.Effect;
+import battle.effect.EffectInterfaces.SapHealthEffect;
 import battle.effect.EffectNamesies;
 import battle.effect.InvokeInterfaces.AlwaysCritEffect;
 import battle.effect.InvokeInterfaces.CritStageEffect;
-import battle.effect.InvokeInterfaces.SapHealthEffect;
 import battle.effect.attack.SelfHealingMove;
 import battle.effect.battle.BattleEffect;
 import battle.effect.pokemon.PokemonEffectNamesies;
